# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Eplus v 0.0.1",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/educational-experiences": {
            "get": {
                "summary": "Mostrar experiencias educativas",
                "operationId": "b855d2b6b4ae361d0d9fd348f632e646",
                "responses": {
                    "200": {
                        "description": "Mostrar todos las experiencias educativas."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "summary": "Crear una nueva experiencia educativa",
                "operationId": "44f612494e2a6542c97926e1b58ce0e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nrc",
                                    "name",
                                    "modality",
                                    "educativeProgram"
                                ],
                                "properties": {
                                    "nrc": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "modality": {
                                        "type": "string"
                                    },
                                    "educativeProgram": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experiencia educativa creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/educational-experiences/create": {
            "get": {
                "summary": "Mostrar formulario para crear una nueva experiencia educativa",
                "operationId": "5c07719bacf25ae1fd51f7b3ce5fcad8",
                "responses": {
                    "200": {
                        "description": "Retorna los programas educativos disponibles medidante un arreglo en el modelo.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/educational-experiences/{id}": {
            "get": {
                "summary": "Mostrar una experiencia educativa",
                "operationId": "9aaa5d57b2f23ad040e7f960bcdbf353",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los detalles de una experiencia educativa."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "summary": "Actualizar una experiencia educativa",
                "operationId": "5fabbf3f276e18f832a2d5a299b601cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nrc",
                                    "name",
                                    "modality"
                                ],
                                "properties": {
                                    "nrc": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "modality": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Experiencia educativa actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "summary": "Eliminar una experiencia educativa",
                "operationId": "31a22bac09612f15965d2a5dbe8c50ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Experiencia educativa eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/educational-experiences/{id}/edit": {
            "get": {
                "summary": "Mostrar formulario para editar una experiencia educativa",
                "operationId": "847989b4113da5f62f2b0933441e0ab1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna los detalles de la experiencia educativa seleccionada para su edición."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}